+++ client-agent/src/.env.dev
@@ -6,6 +6,5 @@ API_GATEWAY_PORT=443
 API_GATEWAY_PATH='/md-collector'
 DB_CONNECTION_STRINGS='postgresql://user:password@host.address/db_name'
 QUERIES_FILE='queries.yaml'
-MOCK=false
 IGNORE_CURRENT_TIME=false
 IGNORE_WINSTON_CONSOLE=false
